pyproject.toml
[project]
name = "codelap-edu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=5.28.2",
    "openai>=1.100.2",
    "pygithub>=2.7.0",
    "python-fasthtml>=0.12.24",
]
